Adding format input field for date/time.
Motivation:
When adding a Datestamp the format of the data/time should be
configuration which is it currently not. This commit should add such a
field.

Modifications:
Added a field labed 'Format' and gave it a default value of YYYY-MM-DD
HH:MM.

Result:
When displayed in fh-ngui there is now a an input field for specifying
the format that the date/time should be in.

JIRA:
https://issues.jboss.org/browse/RHMAP-4771