Adding a toggleDateFormat function to show/hide format field
Motivation:
When the Field type "Date Stamp Options" is set to 'Data & Time' we want
to display the format field but not other wise. This commit adds the
abillity to hide/show the format field.

Modifications:
Added a toggleDateFormat that triggers when the Type field changes, as
well as invoked when the view is rendered.

Result:
The format field is shown only when the type of date is 'datetime'

There is a minor bug here that I'm looking into which is that the input
text field for the format is not inline with it lable but under it.